% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chao_N0.R
\name{chao_N0}
\alias{chao_N0}
\title{The Chao estimate of the total number of unseen variants
based on training gene mutation frequencies}
\usage{
chao_N0(counts = NULL, r = NULL, N_r = NULL, m)
}
\arguments{
\item{counts}{vector of counts or frequencies of the observed variants.}

\item{r}{unique frequencies.}

\item{N_r}{frequency of frequency r.}

\item{m}{training cohort size.}
}
\description{
The Chao estimate of the total number of unseen variants
based on training gene mutation frequencies
}
\details{
Calculates Chao (1987) estimate of the total number of
unseen variants. Also provides an approximate standard error ("se") of the
estimate as an attribute, computed using the formula provided in Chao (1987).

\dontrun{
# load tcga data
data("tcga")
tcga <- data.table::setDT(tcga)

# calculate variant frequencies
var_freq <- tcga[,
                 .(v_f = length(unique(patient_id))),
                 by = .(Hugo_Symbol, Variant)
                 ]

# calculate cohort size
m <- length(unique(tcga$patient_id))


# SGT estimate for t = 0.5, 1, 10
chao_N0(counts = var_freq$v_f, m = m, t = 0.5)
}
}
