% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgt.R
\name{sgt}
\alias{sgt}
\title{Smooth Good-Toulmin estimate of \eqn{\Delta(t)}, the (expected)
number of new variants in
a future (test) cohort that is \eqn{t} times as large as the
training cohort}
\usage{
sgt(counts = NULL, r = NULL, N_r = NULL, m, t = 1, adj = TRUE)
}
\arguments{
\item{counts}{vector of counts or frequencies of the observed variants.}

\item{r}{unique frequencies.}

\item{N_r}{frequency of frequency r.}

\item{m}{training cohort size.}

\item{t}{positive scalar. The proportion of the future (test) cohort size to the
trianing cohort size.}

\item{adj}{logical. Should the Orlitsky et al. adjustment be used?
Defaults to \code{TRUE}. Ignored if \code{t < 1}.}
}
\description{
Smooth Good-Toulmin estimate of \eqn{\Delta(t)}, the (expected)
number of new variants in
a future (test) cohort that is \eqn{t} times as large as the
training cohort
}
\details{
Computes the original Good Toulmin (1956) estimate of \eqn{\Delta(t)} if \code{t <= 1}. If
\code{t > 1}, the Efron-Thisted estimate (if \code{adj = FALSE}) or the
Efron-Thisted estimate with Orlitsky et al. (2016) adjustment (if \code{adj = TRUE}) is computed.
}
\note{
Either (a) \code{counts}, or (b) \code{r} and
\code{N_r} must be provided.
}
\examples{
\dontrun{
# load tcga data
data("tcga")
tcga <- data.table::setDT(tcga)

# calculate variant frequencies
var_freq <- tcga[,
                 .(v_f = length(unique(patient_id))),
                 by = .(Hugo_Symbol, HGVSp_Short)
                 ]

# calculate cohort size
m <- length(unique(tcga$patient_id))


# SGT estimate for t = 0.5, 1, 10
sgt(counts = var_freq$v_f, m = m, t = 0.5)
sgt(counts = var_freq$v_f, m = m, t = 1)
sgt(counts = var_freq$v_f, m = m, t = 10)
}

}
\references{
Good, I. J., & Toulmin, G. H. (1956). The number of
new species, and the increase in population coverage,
when a sample is increased. Biometrika, 43(1–2), 45–63.
https://doi.org/10.1093/biomet/43.1-2.45.

Efron, B., & Thisted, R. (1976). Estimating the Number
of Unsen Species: How Many Words Did Shakespeare
Know? Biometrika, 63(3), 435–447.
Retrieved from http://www.jstor.org/stable/2335721.

Orlitsky, A., Suresh, A. T., & Wu, Y. (2016). Optimal
prediction of the number of unseen species. Proceedings
of the National Academy of Sciences, 113(47), 13283–13288.
https://doi.org/10.1073/pnas.1607774113
}
