% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goodturing_probs.R
\name{goodturing_probs}
\alias{goodturing_probs}
\title{Computing Good Turing probabilities of encountering gene variants
(including hitherto unobserved variants) based on training gene
mutation frequencies}
\usage{
goodturing_probs(
  counts = NULL,
  r = NULL,
  N_r = NULL,
  m = NULL,
  conf = 1.96,
  N0min = 0,
  N0 = NULL,
  N12_imp = 1,
  N = NULL
)
}
\arguments{
\item{counts}{vector of counts or frequencies of the observed variants.}

\item{r}{unique frequencies.}

\item{N_r}{frequency of frequency r.}

\item{m}{training cohort size.}

\item{N0min}{the minimum value of N0, if known, to be used while
estimating \code{N0}. Ignored if N0 is not \code{NULL}.}

\item{N0}{the total number of unobserved variants. If \code{NULL},
\code{N0} is estimated using Chao's formula.}

\item{N12_imp}{imputed value of N1 and N2 if either of them is 0.
Defaults to 1.}

\item{N}{The total number of variants, which is
\code{sum(Nr)} for \code{r >= 0}. Used for computation of N0. Ignored
if N0 is provided.}
}
\description{
Computing Good Turing probabilities of encountering gene variants
(including hitherto unobserved variants) based on training gene
mutation frequencies
}
\examples{
\dontrun{
# load tcga data
data("tcga")
tcga <- data.table::setDT(tcga)

# calculate variant frequencies for KRAS
var_freq <- tcga[Hugo_Symbol == "KRAS",
                 .(v_f = length(unique(patient_id))),
                 by = .(Hugo_Symbol, Variant)
                 ]
v_f <- var_freq$v_f
names(v_f) <- var_freq$Variant

# calculate cohort size
m <- length(unique(tcga$patient_id))


# Good Turing estimates
goodturing_probs(counts = v_f, m = m)
}
}
