% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoodTuringProbs.R
\name{GoodTuringProbs}
\alias{GoodTuringProbs}
\title{computing good turing probabilities}
\usage{
GoodTuringProbs(counts = NULL, r = NULL, N_r = NULL, m = NULL,
  conf = 1.96, N0min = 0)
}
\description{
computing good turing probabilities
}
